Fix circular reading bug fith firmwares > 64 KB

+++ cc2530prog.c
@@ -152,7 +152,7 @@ const uint8_t dma_desc[32] = {
 	0x42                            /* increment source */
 };
 
-static uint16_t flash_ptr = 0;
+static uint32_t flash_ptr = 0;
 
 static void init_flash_ptr(void)
 {
